Rename from "LinuxContainers" to "RemoteImages" to be more agnostic
