+++ src/classes/Controllers/Images/SearchRemoteImagesController.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace dhope0000\LXDClient\Controllers\Images;
+
+use dhope0000\LXDClient\Tools\Images\SearchRemoteImages;
+
+class SearchRemoteImagesController
+{
+    public function __construct(SearchRemoteImages $searchRemoteImages)
+    {
+        $this->getImages = $searchRemoteImages;
+    }
+
+    public function get($urlKey, $searchType, $searchArch)
+    {
+        return $this->getImages->get($urlKey, $searchType, $searchArch);
+    }
+}
