+++ src/tools/generateRecordedActionsObj.php
@@ -74,6 +74,12 @@ foreach ($fqcns as $class) {
         }
     }
 }
+
+// Old routes we just manually add because they wont be found by the script
+// anymore
+$routeNames['dhope0000\\LXDClient\\Controllers\\Images\\ImportLinuxContainersByAliasController\\import'] = 'Import LinunxContainer.Org Image';
+
+
 $routesString = var_export($routeNames, true);
 $classString = str_replace("REPLACE_ME", $routesString, $classString);
 file_put_contents(__DIR__ . "/../classes/Objects/RouteToNameMapping.php", $classString);
